+++ docs/changes.rst
@@ -5,6 +5,7 @@ Not yet released.
 
 * Fixed tracking suggestions in history.
 * Fixed parsing reverse proxy info from Cloudflare.
+* Make parse error lock a component from translating.
 
 `All changes in detail <https://github.com/WeblateOrg/weblate/milestone/85?closed=1>`__.